Update:Adjust BackStage Home Controller Login Action and its View
