+++ WebApplication1/Areas/BackStage/Views/Home/Login.cshtml
@@ -4,7 +4,7 @@
     ViewBag.Title = "後台管理後台";
 }
 
-<h2>登入與註冊</h2>
+<h2>後台管理系統登入</h2>
 
 
 @using (Html.BeginForm())
@@ -33,17 +33,16 @@
         </div>
 
 
-
-        <div class="form-group">
+        <div class="form-group my-3">
             <div class="col-md-offset-2 col-md-10">
-                <input type="submit" value="登入" class="btn btn-default" />
+                <input type="submit" value="登入" class="btn btn-primary" />
             </div>
         </div>
     </div>
 }
 
 <div>
-    @Html.ActionLink("Back to List", "Index")
+    <a href="/Home/Index">返回前台首頁</a>
 </div>
 
 @section Scripts {
