+++ WebApplication1/Areas/BackStage/Views/Shared/_Layout.cshtml
@@ -17,7 +17,7 @@
                     <span class="icon-bar"></span>
                 </button>*@
 
-                @Html.ActionLink("國際縱火調查人員協會臺灣分會-後台管理系統", "Index", "Home", new { area = "" }, new { @class = "navbar-brand btn btn-primary text-white" })
+                @Html.ActionLink("火箭隊後端專題實作-後台管理系統", "Index", "Home", new { area = "BackStage" }, new { @class = "navbar-brand btn btn-primary text-white" })
             </div>
             <div class="navbar-collapse collapse">
                 <ul class="nav navbar-nav">
@@ -30,7 +30,7 @@
         @RenderBody()
         <hr />
         <footer>
-            <p>&copy; @DateTime.Now.Year - 國際縱火調查人員協會臺灣分會-後台管理系統</p>
+            <p>&copy; @DateTime.Now.Year - 火箭隊後端專題實作-後台管理系統</p>
         </footer>
     </div>
