+++ WebApplication1/Utility/Encrypt.cs
@@ -1,0 +1,42 @@
+﻿using Konscious.Security.Cryptography;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Security.Cryptography;
+using System.Text;
+using System.Web;
+
+namespace WebApplication1.Utility
+{
+    public class Encrypt
+    {
+        // Argon2 加密
+        //產生 Salt 功能
+        //使用加密安全隨機數產生器 ( ) 產生隨機鹽
+        public static byte[] CreateSalt()
+        {
+            //建立一個大小為 16 ( ) 的位元組陣列buffer儲存產生的 salt
+            var buffer = new byte[16];
+            var rng = new RNGCryptoServiceProvider();
+            rng.GetBytes(buffer);
+            return buffer;
+        }
+
+        // Hash 處理加鹽的密碼功能
+        //將使用者的密碼和產生的鹽作為輸入，使用 Argon2 演算法執行密碼雜湊
+        public static byte[] HashPassword(string password, byte[] salt)
+        {
+            var argon2 = new Argon2id(Encoding.UTF8.GetBytes(password));
+
+            //底下這些數字會影響運算時間，而且驗證時要用一樣的值
+            //設定之前生成的鹽
+            argon2.Salt = salt;
+            argon2.DegreeOfParallelism = 8; // 4 核心就設成 8
+            argon2.Iterations = 2; // 迭代運算次數，更高的迭代次數可以提高安全性
+            argon2.MemorySize = 1024; // 1 GB，定義演算法要使用的記憶體大小（以位元組為單位）
+
+            //Argon2 演算法產生 16 位元組雜湊並將其作為位元組數組傳回
+            return argon2.GetBytes(16);
+        }
+    }
+}
