+++ WebApplication1/Views/Calendar/Index.cshtml
@@ -1,5 +1,4 @@
-﻿
-@{
+﻿@{
     ViewBag.Title = "Calendar";
     ViewBag.PageTitle = "協會行事曆";
     Layout = "~/Views/Shared/_LayoutPage.cshtml";
