+++ WebApplication1/Views/Contact/Index.cshtml
@@ -21,7 +21,12 @@
 
     <div class="form-horizontal">
 
-        @Html.ValidationSummary(true, "", new { @class = "text-danger" })
+        <div class="form-group">
+            <div class="col-sm-2"></div>
+            <div class="col-md-10">
+                @Html.ValidationSummary(true, "", new { @class = "text-danger" })
+            </div>
+        </div>
 
         <div class="form-group">
             @Html.Label("姓　名：", htmlAttributes: new { @class = "col-sm-2 control-label" })
@@ -107,93 +112,6 @@
         </div>
     </div>
 }
-
-@*<form class="form-horizontal" role="form">
-        @Html.ValidationSummary(true, "", new { @class = "text-danger" })
-        <div class="form-group">
-            <label class="col-sm-2 control-label">姓　名：</label>
-            <div class="col-sm-7">
-                <input class="form-control" placeholder="請輸入姓名">
-            </div>
-            <div class="col-sm-3">
-                <span class="alertred"><i class="fa fa-exclamation-circle"></i> 姓名必填</span>
-            </div>
-        </div>
-
-        <div class="form-group">
-            <label class="col-sm-2 control-label">性　別：</label>
-            <div class="col-sm-7">
-                <label class="checkbox-inline checkbox01">
-                    <input class=" radiostyle01" type="radio" name="radio" id="radio" value="radio" /> 男
-                </label>
-                <label class="checkbox-inline checkbox01">
-                    <input class="radiostyle01" type="radio" name="radio" id="radio" value="radio" /> 女
-                </label>
-            </div>
-            <div class="col-sm-3">
-                <span class="alertred"><i class="fa fa-exclamation-circle"></i> 性別必填</span>
-            </div>
-        </div>
-
-        <div class="form-group">
-            <label class="col-sm-2 control-label">聯絡電話：</label>
-            <div class="col-sm-7">
-                <input class="form-control" placeholder="請輸入聯絡電話">
-            </div>
-            <div class="col-sm-3">
-                <span class="alertred"><i class="fa fa-exclamation-circle"></i> 聯絡電話必填</span>
-            </div>
-        </div>
-
-        <div class="form-group">
-            <label class="col-sm-2 control-label">E-mail：</label>
-            <div class="col-sm-7">
-                <input class="form-control" placeholder="請輸入E-mail">
-            </div>
-            <div class="col-sm-3">
-                <span class="alertred"><i class="fa fa-exclamation-circle"></i> E-mail必填</span>
-            </div>
-        </div>
-
-        <div class="form-group">
-            <label class="col-sm-2 control-label">詢問內容：</label>
-            <div class="col-sm-10">
-                <textarea class="form-control form-control-message" placeholder="請輸入詢問內容" rows="10" required></textarea>
-            </div>
-        </div>
-
-        <div class="form-group">
-            <label class="col-sm-2 control-label">驗證碼：</label>
-            <div class="col-sm-3">
-                <input class="form-control" placeholder="請輸入驗證碼">
-            </div>
-            <div class="col-sm-4">
-                <img src="images/test.jpg" alt="">
-            </div>
-            <div class="col-sm-3">
-                <span class="alertred"><i class="fa fa-exclamation-circle"></i> 驗證碼必填</span>
-            </div>
-        </div>
-
-        <hr>
-
-        <div class="form-group">
-            <div class="col-sm-offset-2 col-sm-10">
-                <button class="btn btn-primary solid blank" type="submit">送 出</button>
-            </div>
-        </div>
-    </form>*@
-
-@*<div class="tags-area clearfix">
-        <div class="share-items pull-right">
-            <ul class="post-social-icons unstyled">
-                <li><a href="#"><i class="fa fa-facebook"></i></a></li>
-                <li><a href="#"><i class="fa fa-twitter"></i></a></li>
-                <li><a href="#"><i class="fa fa-google-plus"></i></a></li>
-            </ul>
-        </div>
-    </div>*@
-
 
 @section  scripts
 {
